requirements.txt
transformers==4.31.0
torch==2.0.1
fastapi==0.101.0
accelerate==0.21.0
bitsandbytes==0.41.1
scipy==1.11.1
uvicorn==0.23.2
openai==0.27.8
python-dotenv==1.0.0
